Add files via upload
First, run pdf.py on the data to find the peak in the curve as the classification threshold; then run biaozhuncha.py to classify the photon divergence method; finally, run shaixuan.py to correct the water photons.
The data processed by this code is the signal data that has been denoised. The denoising result will affect the subsequent data processing, so a better denoising method should be selected. In addition, if some parameters in this code are adjusted according to the characteristics of your own data, better results may be achieved.